Added in measure objects
In pure abjad/lilypond, it's possible for notes to be *longer* than the measure they contain (i.e. a whole note in a 2/4 bar), so this new Measure class prevents that from happening. It trims long notes, then returns how much it trimmed so then hypothetically we can program it so that we get ties over bars, but all in due time